Resolve #782 -- ClientID corrected
* Fixed the whole system of PlayerId and ClientId
* Changed PlayerId (also called UserId) to be ulong
* Added ClientId to be used where it needed

Refs #782 #778

+++ GameServerLib/Chatbox/Commands/CoordsCommand.cs
@@ -21,7 +21,7 @@ namespace LeagueSandbox.GameServer.Chatbox.Commands
 
         public override void Execute(int userId, bool hasReceivedArguments, string arguments = "")
         {
-            var champion = _playerManager.GetPeerInfo(userId).Champion;
+            var champion = _playerManager.GetPeerInfo((ulong)userId).Champion;
             _logger.Debug($"At {champion.X}; {champion.Y}");
             var msg = $"At Coords - X: {champion.X} Y: {champion.Y} Z: {champion.GetZ()}";
             ChatCommandManager.SendDebugMsgFormatted(DebugMsgType.NORMAL, msg);
